requirements.txt
aiocsv==1.3.2
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
bitarray==2.9.2
certifi==2024.6.2
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
eth-account==0.13.0
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
Faker==24.1.0
frozenlist==1.4.1
hexbytes==1.2.1
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.0
lru-dict==1.3.0
multidict==6.0.5
numpy==1.26.4
parsimonious==0.10.0
prompt_toolkit==3.0.47
protobuf==5.27.2
pycryptodome==3.19.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==15.1.0
questionary==1.10.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.1
six==1.16.0
toolz==0.12.1
tqdm==4.66.1
typing_extensions==4.12.2
uint==0.2.1
urllib3==2.2.2
wcwidth==0.2.13
web3==6.10.0
websockets==12.0
yarl==1.9.4
